refactor(ProductService): ProductService 와 ProductProvider 인터페이스 분리
