+++ src/main/java/shopping/product/application/ProductProvider.java
@@ -1,0 +1,7 @@
+package shopping.product.application;
+
+import shopping.product.domain.Product;
+
+public interface ProductProvider {
+    Product findProductById(Long id);
+}
